Wrote basic DataSearch
-Reads and analyses CSVs from factfinder.census.gov
-Generates households based on CSVs (NOTE: based on just number. There
is no age method yet!)
-People befriend everyone in their household
-Generates friendships in the overall network using our old methods
-Does not yet save/open graph, but does save the results of the
simulation to an infojungstorage
Note: This assumes that if you live with someone, you have the same
chance of catching a disease from them as from any of your friends. Age
and other cool thing are also not implemented. This is just a basic
sketch of what this program should be by Monday.
By the way, usually, don't just merge into the master branch like this.